Move image widget classes into widgets.image package
