+++ collect_app/src/main/java/org/odk/collect/android/widgets/image/ImageWidgetContent.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.widgets
+package org.odk.collect.android.widgets.image
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -9,6 +9,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import org.javarosa.form.api.FormEntryPrompt
+import org.odk.collect.android.widgets.MediaWidgetAnswerViewModel
+import org.odk.collect.android.widgets.WidgetAnswer
+import org.odk.collect.android.widgets.WidgetIconButton
 import org.odk.collect.androidshared.ui.compose.marginStandard
 import org.odk.collect.strings.R.string
