+++ collect_app/src/test/java/org/odk/collect/android/widgets/image/ImageWidgetTest.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.widgets
+package org.odk.collect.android.widgets.image
 
 import android.content.Intent
 import android.provider.MediaStore
@@ -25,10 +25,11 @@ import org.mockito.kotlin.whenever
 import org.odk.collect.android.formentry.questions.QuestionDetails
 import org.odk.collect.android.support.MockFormEntryPromptBuilder
 import org.odk.collect.android.support.WidgetTestActivity
+import org.odk.collect.android.widgets.MediaWidgetAnswerViewModel
+import org.odk.collect.android.widgets.QuestionWidget
 import org.odk.collect.android.widgets.base.FileWidgetTest
 import org.odk.collect.android.widgets.support.FakeQuestionMediaManager
 import org.odk.collect.android.widgets.support.FakeWaitingForDataRegistry
-import org.odk.collect.android.widgets.utilities.FileAnswerDelegate
 import org.odk.collect.androidtest.onNodeWithClickLabel
 import org.odk.collect.shared.TempFiles
 import org.odk.collect.strings.R.string
